@@ -1,20 +1,21 @@
-use futures::stream::Stream;
-use futures::StreamExt;
 use opentelemetry::global::shutdown_tracer_provider;
-use opentelemetry::sdk::{
-    export::metrics::{Aggregator, AggregatorSelector, ExportKind, ExportKindFor},
-    metrics::{aggregators, PushController},
+use opentelemetry::runtime;
+use opentelemetry::sdk::export::metrics::aggregation::{
+    AggregationKind, Temporality, TemporalitySelector,
 };
+use opentelemetry::sdk::metrics::aggregators::Aggregator;
+use opentelemetry::sdk::metrics::controllers::BasicController;
+use opentelemetry::sdk::metrics::sdk_api::Descriptor;
+use opentelemetry::sdk::{export::metrics::AggregatorSelector, metrics::aggregators};
 use opentelemetry::trace::TraceError;
-use opentelemetry::{
-    baggage::BaggageExt,
-    metrics::{self, Descriptor, ObserverResult},
-    trace::{TraceContextExt, Tracer},
-    Context, Key, KeyValue,
-};
 use opentelemetry::{
     global,
     sdk::{propagation::TraceContextPropagator, trace as sdktrace, Resource},
+};
+use opentelemetry::{
+    metrics,
+    trace::{TraceContextExt, Tracer},
+    Context, Key, KeyValue,
 };
 use opentelemetry_dynatrace::transform::DimensionSet;
 use opentelemetry_dynatrace::ExportConfig;
@@ -50,14 +51,13 @@ fn init_tracer() -> Result<sdktrace::Tracer, TraceError> {
         .install_batch(opentelemetry::runtime::Tokio)
 }
 
-// Skip first immediate tick from tokio, not needed for async_std.
-fn delayed_interval(duration: Duration) -> impl Stream<Item = tokio::time::Instant> {
-    opentelemetry::sdk::util::tokio_interval_stream(duration).skip(1)
-}
-
-fn init_meter() -> metrics::Result<PushController> {
+fn init_metrics() -> metrics::Result<BasicController> {
     opentelemetry_dynatrace::new_pipeline()
-        .metrics(tokio::spawn, delayed_interval)
+        .metrics(
+            CustomAggregator(),
+            CustomTemporalitySelector(),
+            runtime::Tokio,
+        )
         .with_exporter(
             opentelemetry_dynatrace::new_exporter().with_export_config(
                 ExportConfig::default()
@@ -75,12 +75,10 @@ fn init_meter() -> metrics::Result<PushController> {
             KeyValue::new(semcov::resource::SERVICE_NAME, "rust-quickstart"),
             KeyValue::new(semcov::resource::SERVICE_VERSION, env!("CARGO_PKG_VERSION")),
         ]))
-        .with_export_kind(CustomExportKindFor())
-        .with_aggregator_selector(CustomAggregator())
         .build()
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 struct CustomAggregator();
 
 impl AggregatorSelector for CustomAggregator {
@@ -90,26 +88,21 @@ impl AggregatorSelector for CustomAggregator {
     ) -> Option<Arc<(dyn Aggregator + Sync + std::marker::Send + 'static)>> {
         match descriptor.name() {
             "ex.com.one" => Some(Arc::new(aggregators::last_value())),
-            "ex.com.two" => Some(Arc::new(aggregators::histogram(
-                descriptor,
-                &[0.0, 0.5, 1.0, 10.0],
-            ))),
+            "ex.com.two" => Some(Arc::new(aggregators::histogram(&[0.0, 0.5, 1.0, 10.0]))),
             _ => Some(Arc::new(aggregators::sum())),
         }
     }
 }
 
 #[derive(Debug, Clone)]
-struct CustomExportKindFor();
+struct CustomTemporalitySelector();
 
-impl ExportKindFor for CustomExportKindFor {
-    fn export_kind_for(&self, _descriptor: &Descriptor) -> ExportKind {
-        ExportKind::Delta
+impl TemporalitySelector for CustomTemporalitySelector {
+    fn temporality_for(&self, _descriptor: &Descriptor, _kind: &AggregationKind) -> Temporality {
+        Temporality::Delta
     }
 }
 
-const FOO_KEY: Key = Key::from_static_str("ex.com/foo");
-const BAR_KEY: Key = Key::from_static_str("ex.com/bar");
 const LEMONS_KEY: Key = Key::from_static_str("lemons");
 const ANOTHER_KEY: Key = Key::from_static_str("ex.com/another");
 
@@ -128,28 +121,22 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     // matches the containing block, reporting traces and metrics during the whole
     // execution.
     let _init_tracer = init_tracer()?;
-    let _init_meter = init_meter()?;
+    let metrics_controller = init_metrics()?;
+    let cx = Context::new();
 
     let tracer = global::tracer("ex.com/basic");
     let meter = global::meter("ex.com/basic");
 
-    let one_metric_callback =
-        |res: ObserverResult<f64>| res.observe(1.0, COMMON_ATTRIBUTES.as_ref());
-    let _ = meter
-        .f64_value_observer("ex.com.one", one_metric_callback)
-        .with_description("A ValueObserver set to 1.0")
+    let gauge = meter
+        .f64_observable_gauge("ex.com.one")
+        .with_description("A GaugeObserver set to 1.0")
         .init();
+    meter.register_callback(move |cx| gauge.observe(cx, 1.0, COMMON_ATTRIBUTES.as_ref()))?;
 
-    let histogram_two = meter.f64_histogram("ex.com.two").init();
+    let histogram = meter.f64_histogram("ex.com.two").init();
 
     let another_recorder = meter.f64_histogram("ex.com.two").init();
-    another_recorder.record(5.5, COMMON_ATTRIBUTES.as_ref());
-
-    let _baggage =
-        Context::current_with_baggage(vec![FOO_KEY.string("foo1"), BAR_KEY.string("bar1")])
-            .attach();
-
-    let histogram = histogram_two.bind(COMMON_ATTRIBUTES.as_ref());
+    another_recorder.record(&cx, 5.5, COMMON_ATTRIBUTES.as_ref());
 
     tracer.in_span("operation", |cx| {
         let span = cx.span();
@@ -159,20 +146,13 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
         );
         span.set_attribute(ANOTHER_KEY.string("yes"));
 
-        meter.record_batch_with_context(
-            // Note: call-site variables added as context Entries:
-            &Context::current_with_baggage(vec![ANOTHER_KEY.string("xyz")]),
-            COMMON_ATTRIBUTES.as_ref(),
-            vec![histogram_two.measurement(2.0)],
-        );
-
         tracer.in_span("Sub operation...", |cx| {
             let span = cx.span();
             span.set_attribute(LEMONS_KEY.string("five"));
 
             span.add_event("Sub span event", vec![]);
 
-            histogram.record(1.3);
+            histogram.record(&cx, 1.3, &[]);
         });
     });
 
@@ -180,6 +160,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     tokio::time::sleep(Duration::from_secs(60)).await;
 
     shutdown_tracer_provider();
+    metrics_controller.stop(&cx)?;
 
     Ok(())
 }
