@@ -8,8 +8,12 @@ use hyper::{
 };
 use opentelemetry::{
     global,
-    metrics::{BoundCounter, BoundHistogram},
-    KeyValue,
+    metrics::{Counter, Histogram},
+    sdk::{
+        export::metrics::aggregation,
+        metrics::{controllers, processors, selectors},
+    },
+    Context, KeyValue,
 };
 use opentelemetry_prometheus::PrometheusExporter;
 use prometheus::{Encoder, TextEncoder};
@@ -22,13 +26,14 @@ lazy_static! {
 }
 
 async fn serve_req(
+    cx: Context,
     req: Request<Body>,
     state: Arc<AppState>,
 ) -> Result<Response<Body>, hyper::Error> {
     println!("Receiving request at path {}", req.uri());
     let request_start = SystemTime::now();
 
-    state.http_counter.add(1);
+    state.http_counter.add(&cx, 1, &[]);
 
     let response = match (req.method(), req.uri().path()) {
         (&Method::GET, "/metrics") => {
@@ -36,7 +41,7 @@ async fn serve_req(
             let encoder = TextEncoder::new();
             let metric_families = state.exporter.registry().gather();
             encoder.encode(&metric_families, &mut buffer).unwrap();
-            state.http_body_gauge.record(buffer.len() as u64);
+            state.http_body_gauge.record(&cx, buffer.len() as u64, &[]);
 
             Response::builder()
                 .status(200)
@@ -54,22 +59,34 @@ async fn serve_req(
             .unwrap(),
     };
 
-    state
-        .http_req_histogram
-        .record(request_start.elapsed().map_or(0.0, |d| d.as_secs_f64()));
+    state.http_req_histogram.record(
+        &cx,
+        request_start.elapsed().map_or(0.0, |d| d.as_secs_f64()),
+        &[],
+    );
     Ok(response)
 }
 
 struct AppState {
     exporter: PrometheusExporter,
-    http_counter: BoundCounter<u64>,
-    http_body_gauge: BoundHistogram<u64>,
-    http_req_histogram: BoundHistogram<f64>,
+    http_counter: Counter<u64>,
+    http_body_gauge: Histogram<u64>,
+    http_req_histogram: Histogram<f64>,
 }
 
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let exporter = opentelemetry_prometheus::exporter().init();
+    let controller = controllers::basic(
+        processors::factory(
+            selectors::simple::histogram([1.0, 2.0, 5.0, 10.0, 20.0, 50.0]),
+            aggregation::cumulative_temporality_selector(),
+        )
+        .with_memory(true),
+    )
+    .build();
+
+    let exporter = opentelemetry_prometheus::exporter(controller).init();
+    let cx = Context::new();
 
     let meter = global::meter("ex.com/hyper");
     let state = Arc::new(AppState {
@@ -77,28 +94,30 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         http_counter: meter
             .u64_counter("example.http_requests_total")
             .with_description("Total number of HTTP requests made.")
-            .init()
-            .bind(HANDLER_ALL.as_ref()),
+            .init(),
         http_body_gauge: meter
             .u64_histogram("example.http_response_size_bytes")
             .with_description("The metrics HTTP response sizes in bytes.")
-            .init()
-            .bind(HANDLER_ALL.as_ref()),
+            .init(),
         http_req_histogram: meter
             .f64_histogram("example.http_request_duration_seconds")
             .with_description("The HTTP request latencies in seconds.")
-            .init()
-            .bind(HANDLER_ALL.as_ref()),
+            .init(),
     });
 
     // For every connection, we must make a `Service` to handle all
     // incoming HTTP requests on said connection.
     let make_svc = make_service_fn(move |_conn| {
         let state = state.clone();
+        let cx = cx.clone();
         // This is the `Service` that will handle the connection.
         // `service_fn` is a helper to convert a function that
         // returns a Response into a `Service`.
-        async move { Ok::<_, Infallible>(service_fn(move |req| serve_req(req, state.clone()))) }
+        async move {
+            Ok::<_, Infallible>(service_fn(move |req| {
+                serve_req(cx.clone(), req, state.clone())
+            }))
+        }
     });
 
     let addr = ([127, 0, 0, 1], 3000).into();
