@@ -245,7 +245,7 @@ pub enum Value {
 }
 
 /// Wrapper for string-like values
-#[derive(Clone, PartialEq, Hash)]
+#[derive(Clone, PartialEq, Eq, Hash)]
 pub struct StringValue(OtelString);
 
 impl fmt::Debug for StringValue {
