@@ -1,42 +1,25 @@
 //! # OpenTelemetry Metrics API
 
-use std::borrow::Cow;
 use std::result;
 use std::sync::PoisonError;
+use std::{borrow::Cow, sync::Arc};
 use thiserror::Error;
 
-mod async_instrument;
-mod config;
-mod counter;
-mod descriptor;
-mod histogram;
-mod kind;
+mod instruments;
 mod meter;
 pub mod noop;
-mod number;
-mod observer;
-pub mod registry;
-pub mod sdk_api;
-mod sync_instrument;
-mod up_down_counter;
-mod value_recorder;
 
-use crate::ExportError;
-pub use async_instrument::{AsyncRunner, BatchObserverResult, Observation, ObserverResult};
-pub use config::InstrumentConfig;
-pub use counter::{BoundCounter, Counter, CounterBuilder};
-pub use descriptor::Descriptor;
-pub use histogram::{BoundHistogram, Histogram, HistogramBuilder};
-pub use kind::InstrumentKind;
-pub use meter::{Meter, MeterProvider};
-pub use number::{AtomicNumber, Number, NumberKind};
-pub use observer::{
-    BatchObserver, SumObserver, SumObserverBuilder, UpDownSumObserver, UpDownSumObserverBuilder,
-    ValueObserver, ValueObserverBuilder,
+use crate::{Context, ExportError};
+pub use instruments::{
+    counter::{AsyncCounter, Counter, ObservableCounter, SyncCounter},
+    gauge::{AsyncGauge, ObservableGauge},
+    histogram::{Histogram, SyncHistogram},
+    up_down_counter::{
+        AsyncUpDownCounter, ObservableUpDownCounter, SyncUpDownCounter, UpDownCounter,
+    },
+    InstrumentBuilder,
 };
-pub use sync_instrument::Measurement;
-pub use up_down_counter::{BoundUpDownCounter, UpDownCounter, UpDownCounterBuilder};
-pub use value_recorder::{BoundValueRecorder, ValueRecorder, ValueRecorderBuilder};
+pub use meter::{Meter, MeterProvider};
 
 /// A specialized `Result` type for metric operations.
 pub type Result<T> = result::Result<T, MetricsError>;
@@ -90,7 +73,7 @@ impl<T> From<PoisonError<T>> for MetricsError {
 }
 
 /// Units denote underlying data units tracked by `Meter`s.
-#[derive(Clone, Default, Debug, PartialEq, Hash)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Hash)]
 pub struct Unit(Cow<'static, str>);
 
 impl Unit {
@@ -113,4 +96,170 @@ impl AsRef<str> for Unit {
     fn as_ref(&self) -> &str {
         self.0.as_ref()
     }
+}
+
+/// SDK implemented trait for creating instruments
+pub trait InstrumentProvider {
+    /// creates an instrument for recording increasing values.
+    fn u64_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<Counter<u64>> {
+        Ok(Counter::new(Arc::new(noop::NoopSyncInstrument::new())))
+    }
+
+    /// creates an instrument for recording increasing values.
+    fn f64_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<Counter<f64>> {
+        Ok(Counter::new(Arc::new(noop::NoopSyncInstrument::new())))
+    }
+
+    /// creates an instrument for recording increasing values via callback.
+    fn u64_observable_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableCounter<u64>> {
+        Ok(ObservableCounter::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording increasing values via callback.
+    fn f64_observable_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableCounter<f64>> {
+        Ok(ObservableCounter::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording changes of a value.
+    fn i64_up_down_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<UpDownCounter<i64>> {
+        Ok(UpDownCounter::new(
+            Arc::new(noop::NoopSyncInstrument::new()),
+        ))
+    }
+
+    /// creates an instrument for recording changes of a value.
+    fn f64_up_down_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<UpDownCounter<f64>> {
+        Ok(UpDownCounter::new(
+            Arc::new(noop::NoopSyncInstrument::new()),
+        ))
+    }
+
+    /// creates an instrument for recording changes of a value.
+    fn i64_observable_up_down_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableUpDownCounter<i64>> {
+        Ok(ObservableUpDownCounter::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording changes of a value via callback.
+    fn f64_observable_up_down_counter(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableUpDownCounter<f64>> {
+        Ok(ObservableUpDownCounter::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording the current value via callback.
+    fn u64_observable_gauge(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableGauge<u64>> {
+        Ok(ObservableGauge::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording the current value via callback.
+    fn i64_observable_gauge(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableGauge<i64>> {
+        Ok(ObservableGauge::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording the current value via callback.
+    fn f64_observable_gauge(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<ObservableGauge<f64>> {
+        Ok(ObservableGauge::new(Arc::new(
+            noop::NoopAsyncInstrument::new(),
+        )))
+    }
+
+    /// creates an instrument for recording a distribution of values.
+    fn f64_histogram(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<Histogram<f64>> {
+        Ok(Histogram::new(Arc::new(noop::NoopSyncInstrument::new())))
+    }
+
+    /// creates an instrument for recording a distribution of values.
+    fn u64_histogram(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<Histogram<u64>> {
+        Ok(Histogram::new(Arc::new(noop::NoopSyncInstrument::new())))
+    }
+
+    /// creates an instrument for recording a distribution of values.
+    fn i64_histogram(
+        &self,
+        _name: String,
+        _description: Option<String>,
+        _unit: Option<Unit>,
+    ) -> Result<Histogram<i64>> {
+        Ok(Histogram::new(Arc::new(noop::NoopSyncInstrument::new())))
+    }
+
+    /// Captures the function that will be called during data collection.
+    ///
+    /// It is only valid to call `observe` within the scope of the passed function.
+    fn register_callback(&self, callback: Box<dyn Fn(&Context) + Send + Sync>) -> Result<()>;
 }
