@@ -9,13 +9,18 @@ pub mod tonic {
         common::v1::KeyValue,
         metrics::v1::{number_data_point, AggregationTemporality},
     };
-    use opentelemetry::metrics::{Number, NumberKind};
-    use opentelemetry::sdk::export::metrics::ExportKind;
+    use opentelemetry::{
+        metrics::MetricsError,
+        sdk::{
+            export::metrics::aggregation::Temporality,
+            metrics::sdk_api::{Number, NumberKind},
+        },
+    };
 
     use opentelemetry::{Key, Value};
 
-    /// Convert [`Number`](opentelemetry::metrics::NumberKind) to target type based
-    /// on it's [`NumberKind`](opentelemetry::metrics::NumberKind).
+    /// Convert [`Number`](opentelemetry::sdk::metrics::sdk_api::Number) to target type based
+    /// on it's [`NumberKind`](opentelemetry::sdk::metrics::sdk_api::NumberKind).
     pub trait FromNumber {
         fn from_number(number: Number, number_kind: &NumberKind) -> Self;
     }
@@ -40,11 +45,18 @@ pub mod tonic {
         }
     }
 
-    impl From<ExportKind> for AggregationTemporality {
-        fn from(kind: ExportKind) -> Self {
-            match kind {
-                ExportKind::Cumulative => AggregationTemporality::Cumulative,
-                ExportKind::Delta => AggregationTemporality::Delta,
+    impl From<Temporality> for AggregationTemporality {
+        fn from(temporality: Temporality) -> Self {
+            match temporality {
+                Temporality::Cumulative => AggregationTemporality::Cumulative,
+                Temporality::Delta => AggregationTemporality::Delta,
+                other => {
+                    opentelemetry::global::handle_error(MetricsError::Other(format!(
+                        "Unknown temporality {:?}, using default instead.",
+                        other
+                    )));
+                    AggregationTemporality::Cumulative
+                }
             }
         }
     }
