@@ -1,11 +1,38 @@
-use crate::export::metrics::{AggregatorSelector, Processor};
-use crate::metrics::selectors::simple::Selector;
+use std::sync::Arc;
+
+use opentelemetry_api::metrics::Result;
+
+use crate::{
+    export::metrics::{AggregatorSelector, Checkpointer, LockedCheckpointer, Processor},
+    metrics::{aggregators::Aggregator, sdk_api::Descriptor},
+};
 
 #[derive(Debug)]
-pub struct NoopProcessor;
+struct NoopAggregatorSelector;
 
-impl Processor for NoopProcessor {
-    fn aggregation_selector(&self) -> &dyn AggregatorSelector {
-        &Selector::Exact
+impl AggregatorSelector for NoopAggregatorSelector {
+    fn aggregator_for(
+        &self,
+        _descriptor: &Descriptor,
+    ) -> Option<Arc<dyn Aggregator + Send + Sync>> {
+        None
+    }
+}
+
+#[derive(Debug)]
+pub struct NoopCheckpointer;
+
+impl Processor for NoopCheckpointer {
+    fn aggregator_selector(&self) -> &dyn AggregatorSelector {
+        &NoopAggregatorSelector
+    }
+}
+
+impl Checkpointer for NoopCheckpointer {
+    fn checkpoint(
+        &self,
+        _f: &mut dyn FnMut(&mut dyn LockedCheckpointer) -> Result<()>,
+    ) -> Result<()> {
+        Ok(())
     }
 }
